Started working on the database
I created advisor, services, and user models, as well as dummy values to
go with it. I also updated the manage.py file to include 3 new functions
that import the csv data into the database. This is for testing
purposes, and we will not have any pre-existing user data there, except
maybe a dev mode user.